Removed all instances of RGBA
Certain browsers such as the Samsung Internet Browser, do not support RGBA colours, which causes no colours to display. In the current state, both the background does not display (as it is green coded in RGBA) and the text does not display (as it is white and blends in with the default white background), making the site unusable via the Samsung Internet Browser.