Fix for multiple suburb postcodes
A quick commit to update how the longitude and latitude of where you lock in is generated. Hopefully this fixes the problem where if you try to lock in from a suburb that has multiple suburbs it produced an error. Now we get the location of the store and add or minus (randomly chosen) a random number to the coordinates so you should still be in the same suburb as the service station