Fix OzBargain scraper and locking in
This should give a quick fix to the OzBargain scraper that would give false positives if a different sort of deal is posted (i.e. Large coffee for $1). Now we check for "7-Eleven" and "Fuel" in the header.

Also fixed the payload for locking in, and hopefully if there is an error while locking in it doesn't crash.