Fix issue #76. Add credits.
This should fix issue #76 (unable to lock in price with a new account). If you have a brand new account and have never locked in before, you should now be able to lock in.

Added a credits/thank you section with links to each contributors GitHub page. They have helped on things big and small, so thank you all.